PCH and SDL Issue
Fixed an issue where I had made changes to some SDL stuff, but I changed
it so that I wouldn't have to include an alternate SDL file. I also
fixed an issues with the precompiled header. Also added the *.sln file
to the ignore list, as their is only one project anyway.